requirements.txt
langchain 
langchain-core
youtube-transcript-api
langchain-community
langchain-openai

langchain-huggingface
faiss-cpu
tiktoken
python-dotenv
huggingface-hub
fastapi
uvicorn
pydantic
sentence-transformers